+++ src/main/components/CloudFileDialog/CloudFileDialog.tsx
@@ -2,13 +2,13 @@ import { Button } from "@mui/material"
 import { observer } from "mobx-react-lite"
 import { useCallback } from "react"
 import { localized } from "../../../common/localize/localizedString"
-import { useStores } from "../../hooks/useStores"
 import {
   Dialog,
   DialogActions,
   DialogContent,
   DialogTitle,
-} from "../Dialog/Dialog"
+} from "../../../components/Dialog"
+import { useStores } from "../../hooks/useStores"
 import { CloudFileList } from "./CloudFileList"
 
 export const CloudFileDialog = observer(() => {
