+++ src/main/components/PianoRoll/TimeSignatureDialog.tsx
@@ -8,7 +8,7 @@ import {
   DialogActions,
   DialogContent,
   DialogTitle,
-} from "../Dialog/Dialog"
+} from "../../../components/Dialog"
 
 export interface TimeSignatureDialogProps {
   initialNumerator?: number
